+++ src/main/frontend/src/app/mapping/concepts-dialog/concepts-dialog.component.ts
@@ -32,12 +32,13 @@ export class ConceptsDialogComponent {
     public dialogRef : MatDialogRef<ConceptsDialogComponent>,
     @Inject(MAT_DIALOG_DATA) public data : {
       title : string,
-      subtitle: string | null,
+      messages?: [],
       action : string,
       concepts : { [key : ConceptId] : Concept },
       codes : { [key : VocabularyId] : { [key : CodeId] : Code } },
       vocabularies : VocabularyId[],
       allTopics : AllTopics,
+      showConceptSelectors?: boolean,
     }
   ) { }