+++ src/main/frontend/src/app/mapping/concepts/concepts.component.ts
@@ -303,7 +303,7 @@ export class ConceptsComponent implements OnInit {
     }
   }
 
-  async importCsv(file: File, format: string) {
+  async importCsv(file: File, applyFilter: boolean, format: string) {
     if (this.state.stacks.hasUndo()) {
       alert('You cannot undo this operation, please save your mapping before');
       return;
@@ -313,30 +313,30 @@ export class ConceptsComponent implements OnInit {
       return;
     }
     try {
-      let filter = csvFilter(this.info);
+      let filter = applyFilter ? csvFilter(this.info) : null;
       let imported = await firstValueFrom(
         this.api.importCsv(file, [], format, [], filter)
       );
-      console.log('IMPORTED', imported);
-      let msgs: string[] = [];
+      let messages: string[] = [];
+      messages.push("Please confirm the imported codes. They are structured here by the concepts to which they are associated in the mapping.")
+      messages.push("Codes that are associated with these concepts in other coding systems will be disabled.")
       if (imported.warnings.length) {
         let warnings = imported.warnings.map((s) => `${s}. `).join('');
-        msgs.push(`There were problems with the import: ${warnings}.`);
+        messages.push(warnings);
       }
-      msgs.push("Codes that are associated to these concepts in other coding systems will be disabled.")
       this.csvImportFile = null;
-      //this.run.emit(new ops.AddMapping(imported.mapping));
       let mapping = imported.mapping;
       this.dialog
         .open(ConceptsDialogComponent, {
           data: {
-            title: 'Add the imported codes?',
-            subtitle: msgs.join(" "),
+            title: 'Confirm import',
+            messages,
             action: 'Ok',
             concepts: mapping.concepts,
             codes: mapping.codes,
             vocabularies: Object.keys(imported.mapping.vocabularies),
             allTopics: imported.allTopics,
+            showConceptSelectors: false,
           },
         })
         .afterClosed()